% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giniCI-package.R
\name{giniCI-package}
\alias{giniCI-package}
\title{giniCI: Gini-based Composite Indicators}
\description{
An implementation of Gini-based weighting approaches in constructing
composite indicators, providing functionalities for normalization,
aggregation, and ranking comparison.
}
\references{
Ciommi, M., Gigliarano, C., Emili, A., Taralli, S., & Chelli, F. M. (2017).
A new class of composite indicators for measuring well-being at the local
level: An application to the Equitable and Sustainable Well-being (BES) of
the Italian Provinces. \emph{Ecological Indicators}, 76, 281--296.
}
\author{
Viet Duong Nguyen (maintainer) <\href{mailto:viet-duong.nguyen@outlook.it}{viet-duong.nguyen@outlook.it}> \cr
Chiara Gigliarano \cr
Mariateresa Ciommi
}
\keyword{internal}
